feat(data): implement fast count optimization with CountRequest/CountResult pattern
Implements PROP-fast-count-optimization.md:
- Replace Task<int> CountAsync with CountRequest/CountResult contract
- Add CountStrategy enum: Exact, Fast, Optimized
- Add IsEstimate flag to distinguish exact vs approximate counts
- Change all count types from int to long for large dataset support

Provider-specific optimizations:
- PostgreSQL: pg_stat_user_tables.n_live_tup for fast counts
- SQL Server: sys.dm_db_partition_stats for partition statistics
- MongoDB: estimatedDocumentCount() for metadata-based counts
- SQLite/Redis/JSON/InMemory: Always exact counts (no metadata)

Entity API improvements:
- Simple: await Entity.Count (defaults to optimized)
- Explicit: await Entity.Count.Exact(ct), .Fast(ct), .Optimized(ct)
- Filtered: await Entity.Count.Where(predicate, strategy, ct)

Breaking changes:
- CountAsync now returns CountResult instead of int
- All count values are long instead of int
- Repository method signatures updated across all adapters

Test coverage:
- 88+ comprehensive count tests across all adapters
- Coverage: Entity.Count syntax, IsEstimate flags, CountStrategy behavior
- Provider-specific optimization verification and fallback testing
- Edge cases: empty collections, large values, concurrent access

+++ src/Connectors/Web/GraphQl/Types/EntityCollection.cs
@@ -3,5 +3,5 @@ namespace Koan.Web.Connector.GraphQl.Types;
 public sealed class EntityCollection
 {
     public List<EntityView> Items { get; set; } = new();
-    public int TotalCount { get; set; }
+    public long TotalCount { get; set; }
 }
